+++ ui/src/mixins/view.js
@@ -1,4 +1,3 @@
-import { camelize } from 'humps'
 import { get } from 'lodash'
 import { camelizeFieldsName } from '../helpers'
 
@@ -6,6 +5,11 @@ import { NotifyError } from '../plugins'
 
 export default {
   props: {
+    beforeFetch: {
+      default: null,
+      type: Function
+    },
+
     dialog: {
       type: Boolean
     },
@@ -23,10 +27,10 @@ export default {
 
   data () {
     return {
+      cancelBeforeFetch: false,
       errors: {},
       fields: {},
       metadata: {},
-
       isFetching: false
     }
   },
@@ -58,6 +62,22 @@ export default {
       if (redirect) {
         this.$router.replace({ name: redirect })
       }
+    },
+
+    fetchHandler (payload, resolve) {
+      const hasBeforeFetch = typeof this.beforeFetch === 'function'
+
+      if (hasBeforeFetch && !this.cancelBeforeFetch) {
+        return this.beforeFetch({
+          payload,
+          resolve: payload => resolve(payload),
+          done: () => {
+            this.cancelBeforeFetch = true
+          }
+        })
+      }
+
+      resolve()
     },
 
     setErrors (errors = {}) {